+++ modules/analysis/rosbag-image-pipeline/image_dags/ros_image_pipeline.py
@@ -33,27 +33,40 @@ from airflow.utils.dates import days_ago
 from boto3.dynamodb.conditions import Key
 from boto3.session import Session
 from mypy_boto3_batch.client import BatchClient
+from sagemaker import get_execution_role
+from sagemaker.processing import ProcessingInput, ProcessingOutput, Processor
+from sagemaker.pytorch.processing import PyTorchProcessor
+from sagemaker.session import Session
 
-from demo_dags import batch_creation_and_tracking
-from demo_dags.dag_config import (
-    CONTAINER_TIMEOUT,
-    DAG_ROLE,
-    DEPLOYMENT_NAME,
-    DYNAMODB_TABLE,
-    ECR_REPO_NAME,
-    FARGATE_JOB_QUEUE_ARN,
-    FILE_SUFFIX,
-    MAX_NUM_FILES_PER_BATCH,
-    MEMORY,
-    MODULE_NAME,
-    ON_DEMAND_JOB_QUEUE_ARN,
-    PROVIDER,
-    REGION,
-    SPOT_JOB_QUEUE_ARN,
-    SRC_BUCKET,
-    TARGET_BUCKET,
-    VCPU,
-)
+from image_dags import batch_creation_and_tracking
+from image_dags.dag_config import ADDF_MODULE_METADATA, DEPLOYMENT_NAME, MODULE_NAME, REGION
+
+# SET MODULE VARIABLES FOR IMAGE EXTRACTION
+PROVIDER = "FARGATE"  # One of ON_DEMAND, SPOT, FARGATE
+FILE_SUFFIX = ".bag"
+VCPU = "4"
+MEMORY = "16384"
+CONTAINER_TIMEOUT = 300  # Seconds - must be at least 60 seconds
+IMAGE_TOPICS = ["/flir_adk/rgb_front_left/image_raw", "/flir_adk/rgb_front_right/image_raw"]
+DESIRED_ENCODING = "bgr8"
+
+# SET MODULE VARIABLES FOR OBJECT DETECTION
+MODEL = "yolov5s"
+YOLO_INSTANCE_TYPE = "ml.m5.xlarge"
+
+
+# GET MODULE VARIABLES FROM APP.PY AND DEPLOYSPEC
+addf_module_metadata = json.loads(ADDF_MODULE_METADATA)
+
+DAG_ROLE = addf_module_metadata["DagRoleArn"]
+DYNAMODB_TABLE = addf_module_metadata["DynamoDbTableName"]
+ECR_REPO_NAME = addf_module_metadata["EcrRepoName"]
+FARGATE_JOB_QUEUE_ARN = addf_module_metadata["FargateJobQueueArn"]
+ON_DEMAND_JOB_QUEUE_ARN = addf_module_metadata["OnDemandJobQueueArn"]
+SPOT_JOB_QUEUE_ARN = addf_module_metadata["SpotJobQueueArn"]
+TARGET_BUCKET = addf_module_metadata["TargetBucketName"]
+
+account = boto3.client("sts").get_caller_identity().get("Account")
 
 ValueType = TypeVar("ValueType")
 
@@ -93,6 +106,7 @@ def create_batch_of_drives(ti, **kwargs):
     Put them in Dynamo and Assign to a batch if not assigned already
     Add Drives until reaching max files allowed in 1 batch (hard limit of 10k)
     """
+    # Establish AWS API Connections
     sts_client = boto3.client("sts")
     assumed_role_object = sts_client.assume_role(RoleArn=DAG_ROLE, RoleSessionName="AssumeRoleSession1")
     credentials = assumed_role_object["Credentials"]
@@ -109,6 +123,25 @@ def create_batch_of_drives(ti, **kwargs):
         aws_session_token=credentials["SessionToken"],
     )
 
+    # Validate Config
+    drives_to_process = kwargs["dag_run"].conf["drives_to_process"]
+    example_input = {
+        "drives_to_process": {
+            "drive1": {
+                "bucket": "addf-ros-image-demo-raw-bucket-d2be7d299",
+                "prefix": "rosbag-scene-detection/drive1/",
+            },
+            "drive2": {"bucket": "addf-ros-image-demo-raw-bucket-d2be7d29", "prefix": "rosbag-scene-detection/drive2/"},
+        }
+    }
+
+    for k, v in drives_to_process.items():
+        assert isinstance(k, str), f"expecting config to be like {example_input}, received: {drives_to_process}"
+        assert (
+            "bucket" in v.keys() and "prefix" in v.keys()
+        ), f"expecting config to be like {example_input}, received: {drives_to_process}"
+        assert v["prefix"][-1] == "/"
+
     table = dynamodb.Table(DYNAMODB_TABLE)
     batch_id = kwargs["dag_run"].run_id
 
@@ -118,27 +151,18 @@ def create_batch_of_drives(ti, **kwargs):
     )["Count"]
 
     if files_in_batch > 0:
-        print("Batch Id already exists in tracking table - using existing batch")
+        logger.info("Batch Id already exists in tracking table - using existing batch")
         return files_in_batch
 
-    print("New Batch Id - collecting unprocessed drives from S3 and adding to the batch")
-    files_in_batch = 0
-    while True:
-        print(f"Segments in Batch: {files_in_batch}")
-        next_continuation = None
-        (files_in_batch, next_continuation,) = batch_creation_and_tracking.add_drives_to_batch(
-            table,
-            SRC_BUCKET,
-            MAX_NUM_FILES_PER_BATCH,
-            batch_id,
-            files_in_batch=files_in_batch,
-            next_continuation=None if files_in_batch == 0 else next_continuation,
-            file_suffix=FILE_SUFFIX,
-            s3_client=s3_client,
-        )
-        if next_continuation is None or files_in_batch > MAX_NUM_FILES_PER_BATCH:
-            print(f"Final Batch Size = {files_in_batch} files")
-            return files_in_batch
+    logger.info("New Batch Id - collecting unprocessed drives from S3 and adding to the batch")
+    files_in_batch = batch_creation_and_tracking.add_drives_to_batch(
+        table=table,
+        drives_to_process=drives_to_process,
+        batch_id=batch_id,
+        file_suffix=FILE_SUFFIX,
+        s3_client=s3_client,
+    )
+    return files_in_batch
 
 
 def get_job_queue_name() -> str:
@@ -150,7 +174,7 @@ def get_job_queue_name() -> str:
 
 def get_job_name() -> str:
     v = "".join(random.choice(string.ascii_lowercase) for i in range(6))
-    return f"batch_simple_mock-{v}"
+    return f"ros-image-pipeline-{v}"
 
 
 def get_job_def_name() -> str:
@@ -175,11 +199,9 @@ def get_batch_client() -> BatchClient:
 
 def register_job_definition_on_demand(ti, job_def_name: str) -> str:
     client = get_batch_client()
-    region = REGION
-    account = boto3.client("sts").get_caller_identity().get("Account")
 
     container_properties = {
-        "image": f"{account}.dkr.ecr.{region}.amazonaws.com/{ECR_REPO_NAME}:latest",
+        "image": f"{account}.dkr.ecr.{REGION}.amazonaws.com/{ECR_REPO_NAME}:rostopng",
         "jobRoleArn": DAG_ROLE,
         "environment": [
             {
@@ -249,7 +271,7 @@ with DAG(
         op_kwargs={"job_def_name": job_definition_name},
     )
 
-    def my_func(**kwargs):
+    def batch_operation(**kwargs):
         ti = kwargs["ti"]
         ds = kwargs["ds"]
         array_size = ti.xcom_pull(task_ids="create_batch_of_drives_task", key="return_value")
@@ -272,26 +294,38 @@ with DAG(
                     textwrap.dedent(
                         """\
                             #/usr/bin/env bash
-                            echo "[$(date)] Start $JOB_NAME for batch $BATCH_ID and index: $AWS_BATCH_JOB_ARRAY_INDEX"
-
-                            echo "Getting file from Dynamo for batch $BATCH_ID and index: $AWS_BATCH_JOB_ARRAY_INDEX"
-
-                            python processing_mock/processing.py --tablename $TABLE_NAME \
-                                --s3bucketout $TARGET_BUCKET \
-                                --index $AWS_BATCH_JOB_ARRAY_INDEX --batchid $BATCH_ID
-
+                            echo "Starting ROS"
+                            source /opt/ros/noetic/setup.bash
+                            export PYTHONPATH=$PYTHONPATH:$ROS_PACKAGE_PATH
+                            env
+                            
+                            echo "[$(date)] Start Image Extraction - batch $BATCH_ID, index: $AWS_BATCH_JOB_ARRAY_INDEX"
+                            python3 rostopng/main.py \
+                                --tablename $TABLE_NAME \
+                                --index $AWS_BATCH_JOB_ARRAY_INDEX \
+                                --batchid $BATCH_ID \
+                                --localbagpath $LOCAL_BAG_PATH \
+                                --localimagespath $LOCAL_IMAGES_PATH \
+                                --imagetopics $IMAGE_TOPICS \
+                                --desiredencoding $DESIRED_ENCODING \
+                                --targetbucket $TARGET_BUCKET
                             """
                     ),
                 ],
                 "environment": [
-                    {"name": "BATCH_ID", "value": batch_id},
                     {"name": "TABLE_NAME", "value": DYNAMODB_TABLE},
+                    {"name": "BATCH_ID", "value": batch_id},
                     {"name": "JOB_NAME", "value": batch_id},
                     {"name": "DEBUG", "value": "true"},
                     {
                         "name": "AWS_ACCOUNT_ID",
                         "value": boto3.client("sts").get_caller_identity().get("Account"),
                     },
+                    {"name": "LOCAL_BAG_PATH", "value": "/tmp/input.bag"},
+                    {"name": "LOCAL_IMAGES_PATH", "value": "/tmp/images/"},
+                    {"name": "LOCAL_VIDEO_PATH", "value": "/tmp/output.mp4"},
+                    {"name": "IMAGE_TOPICS", "value": json.dumps(IMAGE_TOPICS)},
+                    {"name": "DESIRED_ENCODING", "value": DESIRED_ENCODING},
                     {"name": "TARGET_BUCKET", "value": TARGET_BUCKET},
                 ],
             },
@@ -299,7 +333,79 @@ with DAG(
 
         op.execute(ds)
 
-    submit_batch_job = PythonOperator(task_id="submit_batch_job", python_callable=my_func)
+    submit_batch_job = PythonOperator(task_id="submit_batch_job", python_callable=batch_operation)
+
+    def sagemaker_yolo_operation(**kwargs):
+
+        # Establish AWS API Connections
+        sts_client = boto3.client("sts")
+        assumed_role_object = sts_client.assume_role(RoleArn=DAG_ROLE, RoleSessionName="AssumeRoleSession1")
+        credentials = assumed_role_object["Credentials"]
+        dynamodb = boto3.resource(
+            "dynamodb",
+            aws_access_key_id=credentials["AccessKeyId"],
+            aws_secret_access_key=credentials["SecretAccessKey"],
+            aws_session_token=credentials["SessionToken"],
+        )
+        table = dynamodb.Table(DYNAMODB_TABLE)
+        batch_id = kwargs["dag_run"].run_id
+
+        # Get Image Directories per Recording File to Label
+        image_directory_items = table.query(
+            KeyConditionExpression=Key("pk").eq(batch_id),
+            Select="SPECIFIC_ATTRIBUTES",
+            ProjectionExpression="raw_image_dirs",
+        )["Items"]
+
+        image_directories = []
+        for item in image_directory_items:
+            image_directories += item["raw_image_dirs"]
+
+        logger.info(f"Starting object detection job for {len(image_directories)} directories")
+
+        processor = Processor(
+            image_uri=f"{account}.dkr.ecr.{REGION}.amazonaws.com/{ECR_REPO_NAME}:yolo",
+            role=get_execution_role(),
+            instance_count=1,
+            instance_type=YOLO_INSTANCE_TYPE,
+            base_job_name=f"YOLO",
+        )
+
+        for image_directory in image_directories:
+            logger.info(f"Starting object detection job for {image_directory}")
+            logger.info(
+                "Job details available at: "
+                f"https://{REGION}.console.aws.amazon.com/sagemaker/home?region={REGION}#/processing-jobs"
+            )
+            processor.run(
+                inputs=[
+                    ProcessingInput(
+                        input_name="data",
+                        source=f"s3://{TARGET_BUCKET}/{image_directory}/",
+                        destination="/opt/ml/processing/input/",
+                    )
+                ],
+                outputs=[
+                    ProcessingOutput(
+                        output_name="output",
+                        source="/opt/ml/processing/output/",
+                        destination=f"s3://{TARGET_BUCKET}/{image_directory}_post_obj_dets/",
+                    )
+                ],
+                arguments=["--model", MODEL],
+                wait=False,
+                logs=False,
+            )
+
+        logger.info("Waiting on all jobs to finish")
+        logger.info(f"Jobs: {processor.jobs}")
+        for job in processor.jobs:
+            logger.info(f"Waiting on: {job} - logs from job:")
+            job.wait(logs=True)
+
+        logger.info(f"All object detection jobs complete")
+
+    submit_yolo_job = PythonOperator(task_id="submit_yolo_job", python_callable=sagemaker_yolo_operation)
 
     deregister_batch_job_definition = PythonOperator(
         task_id="deregister_batch_job_definition",
@@ -315,4 +421,5 @@ with DAG(
         >> register_batch_job_defintion
         >> submit_batch_job
         >> deregister_batch_job_definition
+        >> submit_yolo_job
     )